[test] Fixing compiler warnings for tests

+++ test/collective/allgather.cpp
@@ -21,7 +21,7 @@ SCENARIO("gather values into all processes", "[collective][allgather]")
         auto result = mpi::allgather(&value, 1, mpi::world, mpi::flag::uniform_t());
 
         THEN("all processes have all values") {
-            REQUIRE(result.count == mpi::global::size);
+            REQUIRE(result.count == (size_t) mpi::global::size);
             for (int i = 0; i < mpi::global::size; ++i)
                 REQUIRE(result[i] == (i + 1));
         }
@@ -42,7 +42,7 @@ SCENARIO("gather values into all processes", "[collective][allgather]")
         auto result = mpi::allgather(value, mpi::world, mpi::flag::uniform_t());
 
         THEN("all processes have all values") {
-            REQUIRE(result.count == (quantity * mpi::global::size));
+            REQUIRE(result.count == (size_t) (quantity * mpi::global::size));
             for (int i = 0, k = 0; i < mpi::global::size; ++i)
                 for (int j = 0; j < quantity; ++j, ++k)
                     REQUIRE(result[k] == (10 * i + j));
@@ -64,7 +64,7 @@ SCENARIO("gather values into all processes", "[collective][allgather]")
 
         THEN("all processes have all values") {
             auto size = mpi::global::size;
-            REQUIRE(result.count == (size * (size + 1) / 2));
+            REQUIRE(result.count == (size_t) (size * (size + 1) / 2));
             for (int i = 0, k = 0; i < mpi::global::size; ++i)
                 for (int j = 0; j <= i; ++j, ++k)
                     REQUIRE(result[k] == (i * 10 + j));
